Fix theme loading races
